+++ main.go
@@ -5,10 +5,14 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/spf13/viper"
 	"github.com/tphakala/birdnet-go/cmd"
 	"github.com/tphakala/birdnet-go/internal/conf"
 	"github.com/tphakala/birdnet-go/internal/httpcontroller"
 )
+
+// buildTime is the time when the binary was built.
+var buildDate string
 
 //go:embed assets/*
 var assetsFs embed.FS
@@ -26,6 +30,8 @@ func main() {
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Error loading configuration: %v\n", err)
 		os.Exit(1)
+	} else {
+		fmt.Printf("BirdNET-Go build date: %s, using config file: %s\n", buildDate, viper.ConfigFileUsed())
 	}
 
 	// Execute the root command